Module Jose.Header

type t = {

alg : Jwa.alg;

jku : string option;

jwk : Jwk.public Jwk.t option;

kid : string option;

x5t : string option;

x5t256 : string option;

typ : string option;

cty : string option;

enc : Jwa.enc option;

}

The header has the following properties:

alg Jwa - RS256 and none is currently the only supported algs

jku JWK Set URL

jwk JSON Web Key

kid Key ID - We currently always expect this to be there, this can change in the future

x5t X.509 Certificate SHA-1 Thumbprint

x5t#S256 X.509 Certificate SHA-256 Thumbprint

typ Type

cty Content Type Not implemented:

x5u X.509 URL

x5c X.509 Certficate Chain

crit Critical

Link to RFC

val make_header : ?⁠typ:string -> ?⁠alg:Jwa.alg -> ?⁠enc:Jwa.enc -> Jwk.priv Jwk.t -> t

make_header typ alg enc jwk if alg is not provided it will be derived from jwk.

val of_string : string -> (t, [> `Msg of string ]) Stdlib.result

val to_string : t -> string

val to_json : t -> Yojson.Safe.t

val of_json : Yojson.Safe.t -> (t, [> `Msg of string ]) Stdlib.result