Module Jwk.Priv

Private keys

These keys are not safe to show and should be used to sign content.

type t = {

kty : string;

n : string;

e : string;

d : string;

p : string;

q : string;

dp : string;

dq : string;

qi : string;

alg : string option;

kid : string;

}

t describes a Private JSON Web Key

val of_priv : Nocrypto.Rsa.priv -> (t, [ `Msg of string ]) Stdlib.result

of_priv priv takes a private key generated by Nocrypto and returns a result t or a message of what went wrong.

val to_priv : t -> (Nocrypto.Rsa.priv, [ `Msg of string ]) Stdlib.result

to_priv t takes a private JWK and returns a result Nocrypto.Rsa.priv or a message of what went wrong.

val of_priv_pem : string -> (t, [ `Msg of string ]) Stdlib.result

of_priv_pem pem takes a PEM as a string and returns a result a result t or a message of what went wrong.

val to_priv_pem : t -> (string, [ `Msg of string ]) Stdlib.result

to_priv_pem t takes a private JWK and returns a result PEM string or a message of what went wrong.

val to_json : t -> Yojson.Safe.t

to_json t takes a t and returns a Yojson.Safe.t

val of_json : Yojson.Safe.t -> (t, [ `Msg of string ]) Stdlib.result

of_json json takes a Yojson.Safe.t and returns a t

val of_string : string -> (t, [ `Msg of string ]) Stdlib.result

of_string json_string takes a JSON string representation and tries to return a t

val to_string : t -> string

to_string t takes a t and returns a JSON string representation