Module Jwk.Pub

Public keys

These keys are safe to show and should be used to verify signed content.

type oct = {

kty : Jwa.kty;

alg : Jwa.alg;

kid : string;

k : string;

}

oct represents a JWK with kty `oct

type rsa = {

alg : Jwa.alg;

kty : Jwa.kty;

use : string option;

n : string;

e : string;

kid : string;

x5t : string option;

}

rsa represents a JWK with kty `RSA

type t =

| RSA of rsa

| OCT of oct
t describes a Public JSON Web Key

val get_kid : t -> string

get_kid jwk is a convencience function to get the kid string

val get_kty : t -> Jwa.kty

get_kty jwk is a convencience function to get the key type

val get_alg : t -> Jwa.alg

get_alg jwk is a convencience function to get the algorithm

val rsa_of_pub : Nocrypto.Rsa.pub -> (rsa, [ `Msg of string ]) Stdlib.result

rsa_of_pub pub takes a public key generated by Nocrypto and returns a result t or a message of what went wrong.

val rsa_to_pub : rsa -> (Nocrypto.Rsa.pub, [ `Msg of string ]) Stdlib.result

rsa_to_pub t takes a public JWK and returns a result Nocrypto.Rsa.pub or a message of what went wrong.

val rsa_of_pub_pem : string -> (rsa, [ `Msg of string ]) Stdlib.result

rsa_of_pub_pem pem takes a public PEM as a string and returns a result a result t or a message of what went wrong.

val rsa_to_pub_pem : rsa -> (string, [ `Msg of string ]) Stdlib.result

rsa_to_pub_pem t takes a public JWK and returns a result public PEM string or a message of what went wrong.

val oct_of_string : string -> oct

oct_of_string secret creates a oct from a shared secret

val to_json : t -> Yojson.Safe.t

to_json t takes a t and returns a Yojson.Safe.t

val of_json : Yojson.Safe.t -> (t, [ `Msg of string ]) Stdlib.result

of_json json takes a Yojson.Safe.t and returns a t

val of_string : string -> (t, [ `Msg of string ]) Stdlib.result

of_string json_string takes a JSON string representation and tries to return a t

val to_string : t -> string

to_string t takes a t and returns a JSON string representation