Module Jose.Jwe

Link to RFC

type t = {

header : Header.t;

cek : string;
(*
Content Encryption Key
*)

iv : string;
(*
Initialization Vector
*)

payload : string;
(*
plaintext to be encrypted
*)

aad : string option;
(*
Additional Authentication Data, for future use
*)

}

A JWE ready for encryption

val make : header:Header.t -> string -> (t, [> `Missing_enc | `Unsupported_alg ]) Stdlib.result

make header payload creates a JWE from a Header.t and the plaintext that you want to encrypt

val encrypt : jwk:'a Jwk.t -> t -> (string, [> `Invalid_alg | `Missing_enc | `Unsupported_enc | `Unsupported_kty ]) Stdlib.result

encrypt jwk t encrypts a t into the compact string format

val decrypt : jwk:Jwk.priv Jwk.t -> string -> (t, [> `Invalid_JWE | `Invalid_JWK | `Decrypt_cek_failed | `Msg of string ]) Stdlib.result

decrypt jwk string decrypts a compact string formated JWE into a t