Module Jwk.Priv

Private keys

These keys are not safe to show and should be used to sign content.

type oct = {

kty : Jwa.kty;

alg : Jwa.alg;

kid : string;

k : string;

}

type rsa = {

alg : Jwa.alg;

kty : Jwa.kty;

n : string;

e : string;

d : string;

p : string;

q : string;

dp : string;

dq : string;

qi : string;

kid : string;

}

rsa describes a Private RSA JSON Web Key

type t =

| RSA of rsa

| OCT of oct

val get_kid : t -> string

get_kid jwk is a convencience function to get the kid string

val get_kty : t -> Jwa.kty

get_kty jwk is a convencience function to get the key type

val get_alg : t -> Jwa.alg

get_alg jwk is a convencience function to get the algorithm

val rsa_of_priv : Nocrypto.Rsa.priv -> (rsa, [ `Msg of string ]) Stdlib.result

of_priv priv takes a private key generated by Nocrypto and returns a result t or a message of what went wrong.

val rsa_to_priv : rsa -> (Nocrypto.Rsa.priv, [ `Msg of string ]) Stdlib.result

to_priv t takes a private JWK and returns a result Nocrypto.Rsa.priv or a message of what went wrong.

val rsa_of_priv_pem : string -> (rsa, [ `Msg of string ]) Stdlib.result

of_priv_pem pem takes a PEM as a string and returns a result a result t or a message of what went wrong.

val rsa_to_priv_pem : rsa -> (string, [ `Msg of string ]) Stdlib.result

to_priv_pem t takes a private JWK and returns a result PEM string or a message of what went wrong.

val to_json : t -> Yojson.Safe.t

to_json t takes a t and returns a Yojson.Safe.t

val of_json : Yojson.Safe.t -> (t, [ `Msg of string ]) Stdlib.result

of_json json takes a Yojson.Safe.t and returns a t

val of_string : string -> (t, [ `Msg of string ]) Stdlib.result

of_string json_string takes a JSON string representation and tries to return a t

val to_string : t -> string

to_string t takes a t and returns a JSON string representation