Module Jwk.Pub

Public keys

These keys are safe to show and should be used to verify signed content.

type t = {

alg : Algorithm.t;

kty : string;

use : string option;

n : string;

e : string;

kid : string;

x5t : string option;

}

t describes a Public JSON Web Key

val of_pub : Nocrypto.Rsa.pub -> (t, [ `Msg of string ]) Stdlib.result

of_pub pub takes a public key generated by Nocrypto and returns a result t or a message of what went wrong.

val to_pub : t -> (Nocrypto.Rsa.pub, [ `Msg of string ]) Stdlib.result

to_pub t takes a public JWK and returns a result Nocrypto.Rsa.pub or a message of what went wrong.

val of_pub_pem : string -> (t, [ `Msg of string ]) Stdlib.result

of_pub_pem pem takes a public PEM as a string and returns a result a result t or a message of what went wrong.

val to_pub_pem : t -> (string, [ `Msg of string ]) Stdlib.result

to_pub_pem t takes a public JWK and returns a result public PEM string or a message of what went wrong.

val to_json : t -> Yojson.Safe.t

to_json t takes a t and returns a Yojson.Safe.t

val of_json : Yojson.Safe.t -> (t, [ `Msg of string ]) Stdlib.result

of_json json takes a Yojson.Safe.t and returns a t

val of_string : string -> (t, [ `Msg of string ]) Stdlib.result

of_string json_string takes a JSON string representation and tries to return a t

val to_string : t -> string

to_string t takes a t and returns a JSON string representation