Module Jose.Jwks

JSON Web Key Set

Link to RFC

type t = {

keys : Jwk.public Jwk.t list;

}

t describes a Private JSON Web Key Set

val to_json : t -> Yojson.Safe.t

to_json t takes a t and returns a Yojson.Safe.t

val of_json : Yojson.Safe.t -> t

of_json json takes a Yojson.Safe.t and returns a t. Keys that can not be serialized safely will be removed from the list

val of_string : string -> t

of_string json_string takes a JSON string representation and returns a t. Keys that can not be serialized safely will be removed from the list

val to_string : t -> string

to_string t takes a t and returns a JSON string representation

val find_key : t -> string -> Jwk.public Jwk.t option